refactor: InputView, OutputView -> AskView, ResultView
